Added files from local to repository.
Added the following files from local to repository
api.php -- contains the connection to the database and json response
db.sql -- contains the SQL statement for creation of the Table to store signup data. 
script.js -- Contains the angular script for input validation in the signup form and handling json response for successful or failed response upon form submit.
style.css -- Contains basic styling and css for the signup form. 
signup.html -- Contains the html code for the signup form.